chore: fix type
